Mild refactor
