+++ packages/special-pages/pages/special-error/app/components/AdvancedInfo.module.css
@@ -42,7 +42,7 @@
 /* Platform-specific styles */
 
 /* macOS */
-[data-platform="macos"] {
+[data-platform-name="macos"] {
     & .container {
         background: var(--advanced-info-bg);
     }
@@ -53,7 +53,7 @@
 }
 
 /* iOS */
-[data-platform="ios"] {
+[data-platform-name="ios"] {
     & .container {
         align-items: center;
     }
