+++ packages/special-pages/pages/special-error/app/providers/MessagingProvider.js
@@ -2,7 +2,7 @@ import { createContext, h } from 'preact'
 import { useContext } from 'preact/hooks'
 
 export const MessagingContext = createContext({
-    messaging: /** @type {import('../src/js/index').SpecialErrorPage | null} */(null)
+    messaging: /** @type {import('../../src/js/index').SpecialErrorPage | null} */(null)
 })
 
 /**
@@ -10,7 +10,7 @@ export const MessagingContext = createContext({
  *
  * @param {Object} props
  * @param {import("preact").ComponentChild} props.children
- * @param {import('../src/js/index').SpecialErrorPage | null} props.messaging
+ * @param {import('../../src/js/index').SpecialErrorPage | null} props.messaging
  */
 export function MessagingProvider ({ children, messaging }) {
     return (<MessagingContext.Provider value={{ messaging }}>
