+++ packages/special-pages/pages/special-error/src/js/index.js
@@ -9,8 +9,8 @@
 import { createTypedMessages } from '@duckduckgo/messaging'
 import { Environment } from '../../../../shared/environment.js'
 import { createSpecialPageMessaging } from '../../../../shared/create-special-page-messaging.js'
-import { init } from '../../app/index.js'
 import { sampleData } from './sampleData.js'
+import { init } from '../../app/index.js'
 
 export class SpecialErrorPage {
     /**
@@ -35,33 +35,6 @@ export class SpecialErrorPage {
      * @returns {Promise<import('../../../../types/special-error').InitialSetupResponse>}
      */
     initialSetup () {
-        // TODO: Remove all integration rigging
-        if (this.integration) {
-            const searchParams = new URLSearchParams(window.location.search)
-            const errorId = searchParams.get('errorId')
-            const platformName = searchParams.get('platformName')
-
-            /** @type {import('../../../../types/special-error').InitialSetupResponse['errorData']} */
-            let errorData = sampleData['ssl.expired'].data
-            if (errorId && Object.keys(sampleData).includes(errorId)) {
-                errorData = sampleData[errorId].data
-            }
-
-            const supportedPlatforms = ['macos', 'ios']
-            /** @type {import('../../../../types/special-error').InitialSetupResponse['platform']} */
-            let platform = { name: 'macos' }
-            if (platformName && supportedPlatforms.includes(platformName)) {
-                platform = { name: /** @type {import('../../../../types/special-error').InitialSetupResponse['platform']['name']} */(platformName) }
-            }
-
-            return Promise.resolve({
-                env: 'development',
-                locale: 'en',
-                platform,
-                errorData
-            })
-        }
-
         return this.messaging.request('initialSetup')
     }
 
@@ -97,6 +70,37 @@ export class SpecialErrorPage {
     }
 }
 
+export class IntegrationSpecialErrorPage extends SpecialErrorPage {
+    /**
+     * @returns {Promise<import('../../../../types/special-error').InitialSetupResponse>}
+     */
+    initialSetup () {
+        const searchParams = new URLSearchParams(window.location.search)
+        const errorId = searchParams.get('errorId')
+        const platformName = searchParams.get('platformName')
+
+        /** @type {import('../../../../types/special-error').InitialSetupResponse['errorData']} */
+        let errorData = sampleData['ssl.expired'].data
+        if (errorId && Object.keys(sampleData).includes(errorId)) {
+            errorData = sampleData[errorId].data
+        }
+
+        const supportedPlatforms = ['macos', 'ios']
+        /** @type {import('../../../../types/special-error').InitialSetupResponse['platform']} */
+        let platform = { name: 'macos' }
+        if (platformName && supportedPlatforms.includes(platformName)) {
+            platform = { name: /** @type {import('../../../../types/special-error').InitialSetupResponse['platform']['name']} */(platformName) }
+        }
+
+        return Promise.resolve({
+            env: 'development',
+            locale: 'en',
+            platform,
+            errorData
+        })
+    }
+}
+
 const baseEnvironment = new Environment()
     .withPlatform(document.documentElement.dataset.platform)
     .withEnv(import.meta.env)
@@ -107,7 +111,9 @@ const messaging = createSpecialPageMessaging({
     pageName: /** @type {string} */(import.meta.pageName)
 })
 
-const specialErrorPage = new SpecialErrorPage(messaging, baseEnvironment.platform)
+const specialErrorPage = baseEnvironment.platform === 'integration'
+    ? new IntegrationSpecialErrorPage(messaging, baseEnvironment.platform)
+    : new SpecialErrorPage(messaging, baseEnvironment.platform)
 
 init(specialErrorPage, baseEnvironment).catch(e => {
     // messages.
